Remove leftover TypeScript file
