+++ README.md
@@ -4,9 +4,13 @@ A simple command line tool for sending UDP packets to barn lights controllers. T
 
 ## Usage
 
+Run the executable script with Node:
+
 ```bash
-lights-sender [--config <path>] [--log-level <level>]
+node bin/lights-sender [--config <path>] [--log-level <level>]
 ```
+
+If you install the package globally or invoke it with `npx`, you can run `lights-sender` directly.
 
 Options:
 
@@ -21,4 +25,4 @@ Options:
 
 ## Development
 
-The entry point is in `src/cli.ts`. The executable script is `bin/lights-sender` which invokes `main()` from the CLI module.
+The entry point is in `src/cli.js`. The executable script is `bin/lights-sender` which invokes `main()` from the CLI module.
