+++ package.json
@@ -10,10 +10,6 @@
   "author": "",
   "license": "ISC",
   "type": "commonjs",
-  "devDependencies": {
-    "ts-node": "^10.9.2",
-    "typescript": "^5.9.2"
-  },
   "bin": {
     "lights-sender": "bin/lights-sender"
   }
